+++ modules/tinymce/src/core/test/ts/browser/selection/DetailsElementTest.ts
@@ -1,64 +1,68 @@
-import { ApproxStructure } from '@ephox/agar';
 import { describe, it } from '@ephox/bedrock-client';
 import { TinyAssertions, TinyHooks } from '@ephox/wrap-mcagar';
+import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
+
+const createAccordion = (
+  { open = true, summary = 'Accordion summary...', body = '<p>Accordion body...</p>' }:
+  { open?: boolean; summary?: string; body?: string } = {}): string =>
+  `<details class="mce-accordion"${open ? ` open="open"` : ''}><summary class="mce-accordion-summary">${summary}</summary>${body}</details>`;
 
 describe('browser.tinymce.selection.DetailsElementTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
     indent: false,
+    entities: 'raw',
     base_url: '/project/tinymce/js/tinymce'
   }, []);
 
-  it('Should should retain open attribute if it is not opened', () => {
+  it('TINY-9731: Should not modify initial details if details_initial_state == inherited', () => {
     const editor = hook.editor();
-    editor.setContent('<details><summary>a</summary>b</details>');
-    TinyAssertions.assertContentStructure(editor,
-      ApproxStructure.build((s, str, _arr) => s.element('body', {
-        children: [
-          s.element('details', {
-            attrs: {
-              'open': str.is('open'),
-              'data-mce-open': str.none('Should not have a data attr')
-            },
-            children: [
-              s.element('summary', {
-                children: [
-                  s.text(str.is('a'))
-                ]
-              }),
-              s.text(str.is('b'))
-            ]
-          })
-        ]
-      }))
-    );
-    TinyAssertions.assertContent(editor, '<details><summary>a</summary>b</details>');
+    editor.options.set('details_initial_state', 'inherited');
+    const content = [ createAccordion({ open: true }), createAccordion({ open: false }) ].join('');
+    editor.setContent(content);
+    TinyAssertions.assertContent(editor, content);
+    editor.options.unset('details_initial_state');
   });
 
-  it('Should should retain open attribute if it opened', () => {
+  it('TINY-9731: Should expand initial details if details_initial_state == expanded', () => {
     const editor = hook.editor();
-    editor.setContent('<details open="open"><summary>a</summary>b</details>');
-    TinyAssertions.assertContentStructure(editor,
-      ApproxStructure.build((s, str, _arr) => s.element('body', {
-        children: [
-          s.element('details', {
-            attrs: {
-              'open': str.is('open'),
-              'data-mce-open': str.is('open')
-            },
-            children: [
-              s.element('summary', {
-                children: [
-                  s.text(str.is('a'))
-                ]
-              }),
-              s.text(str.is('b'))
-            ]
-          })
-        ]
-      }))
-    );
-    TinyAssertions.assertContent(editor, '<details open="open"><summary>a</summary>b</details>');
+    editor.options.set('details_initial_state', 'expanded');
+    editor.setContent([ createAccordion({ open: true }), createAccordion({ open: false }) ].join(''));
+    TinyAssertions.assertContent(editor, [ createAccordion({ open: true }), createAccordion({ open: true }) ].join(''));
+    editor.options.unset('details_initial_state');
+  });
+
+  it('TINY-9731: Should collapse initial details if details_initial_state == collapsed', () => {
+    const editor = hook.editor();
+    editor.options.set('details_initial_state', 'collapsed');
+    editor.setContent([ createAccordion({ open: true }), createAccordion({ open: false }) ].join(''));
+    TinyAssertions.assertContent(editor, [ createAccordion({ open: false }), createAccordion({ open: false }) ].join(''));
+    editor.options.unset('details_initial_state');
+  });
+
+  it('TINY-9731: Should not modify serialized details if details_serialized_state == inherited', () => {
+    const editor = hook.editor();
+    editor.options.set('details_serialized_state', 'inherited');
+    const content = [ createAccordion({ open: true }), createAccordion({ open: false }) ].join('');
+    editor.setContent(content);
+    assert.equal(content, editor.getContent());
+    editor.options.unset('details_serialized_state');
+  });
+
+  it('TINY-9731: Should expand serialized details if details_serialized_state == expanded', () => {
+    const editor = hook.editor();
+    editor.options.set('details_serialized_state', 'expanded');
+    editor.setContent([ createAccordion({ open: true }), createAccordion({ open: false }) ].join(''));
+    TinyAssertions.assertContent(editor, [ createAccordion({ open: true }), createAccordion({ open: true }) ].join(''));
+    editor.options.unset('details_serialized_state');
+  });
+
+  it('TINY-9731: Should collapse serialized details if details_serialized_state == collapsed', () => {
+    const editor = hook.editor();
+    editor.options.set('details_serialized_state', 'collapsed');
+    editor.setContent([ createAccordion({ open: true }), createAccordion({ open: false }) ].join(''));
+    TinyAssertions.assertContent(editor, [ createAccordion({ open: false }), createAccordion({ open: false }) ].join(''));
+    editor.options.unset('details_serialized_state');
   });
 });